refactor: core unit index
