+++ pages/core-units.tsx
@@ -1,12 +1,12 @@
 import type { CoreUnit } from '@/core/models/interfaces/coreUnit';
-import CoreUnitsView from '@/views/CoreUnits/CoreUnitsView';
-import { fetchCoreUnits } from '@/views/CoreUnits/cuTableAPI';
+import CoreUnitsIndexView from '@/views/CoreUnitsIndexView/CoreUnitsIndexView';
+import { fetchCoreUnits } from '@/views/CoreUnitsIndexView/cuTableAPI';
 import type { GetServerSideProps, NextPage } from 'next';
 
 interface Props {
   coreUnits: CoreUnit[];
 }
-const CuTablePage: NextPage<Props> = ({ coreUnits }) => <CoreUnitsView coreUnits={coreUnits} />;
+const CuTablePage: NextPage<Props> = ({ coreUnits }) => <CoreUnitsIndexView coreUnits={coreUnits} />;
 
 export const getServerSideProps: GetServerSideProps = async () => {
   const coreUnits = await fetchCoreUnits();
