+++ src/stories/components/CoreUnitSummary/CoreUnitSummary.tsx
@@ -3,7 +3,7 @@ import { siteRoutes } from '@ses/config/routes';
 import _ from 'lodash';
 import { useRouter } from 'next/router';
 import React, { forwardRef, useCallback, useMemo } from 'react';
-import { useCoreUnitsTableView } from '@/views/CoreUnits/useCoreUnitsTableView';
+import { useCoreUnitsTableView } from '@/views/CoreUnitsIndexView/useCoreUnitsTableView';
 import { useThemeContext } from '../../../core/context/ThemeContext';
 import { filterData, getArrayParam, getStringParam } from '../../../core/utils/filters';
 import { buildQueryString } from '../../../core/utils/urls';
