+++ src/views/CoreUnitsIndexView/CuTableRenders.tsx
@@ -17,11 +17,11 @@ import {
 import type { TeamStatus } from '@/core/models/interfaces/types';
 import { getShortCode } from '@/core/utils/string';
 import CoreUnitCard from '@/stories/components/CoreUnitCard/CoreUnitCard';
-import { CuRenderLinks } from './CuRenderLinks/CuRenderLinks';
-import { CuTableColumnExpenditures } from './CuTableColumnExpenditures/CuTableColumnExpenditures';
-import { CuTableColumnSummary } from './CuTableColumnSummary/CuTableColumnSummary';
-import CuTableColumnTeamMember from './CuTableColumnTeamMember/CuTableColumnTeamMember';
-import LastModifiedActorCoreUnit from './LastModifiedActorCoreUnit/LastModifiedActorCoreUnit';
+import { CuRenderLinks } from './components/CuRenderLinks/CuRenderLinks';
+import { CuTableColumnExpenditures } from './components/CuTableColumnExpenditures/CuTableColumnExpenditures';
+import { CuTableColumnSummary } from './components/CuTableColumnSummary/CuTableColumnSummary';
+import CuTableColumnTeamMember from './components/CuTableColumnTeamMember/CuTableColumnTeamMember';
+import LastModifiedActorCoreUnit from './components/LastModifiedActorCoreUnit/LastModifiedActorCoreUnit';
 import type { CoreUnit } from '@ses/core/models/interfaces/coreUnit';
 
 export const renderSummary = (coreUnit: CoreUnit) => (
