+++ src/views/CoreUnitsIndexView/components/CuTableColumnExpenditures/CuTableColumnExpendituresSkeleton.tsx
