+++ src/views/CoreUnitsIndexView/components/CuTableColumnSummary/CuTableColumnSummarySkeleton.tsx
