+++ src/views/CoreUnitsIndexView/components/CustomTable/CardItemCoreUnitMobile/CardItemCoreUnitMobile.tsx
