+++ src/views/CoreUnitsIndexView/components/CustomTable/ItemCoreUnit/ItemCoreUnit.tsx
