+++ src/views/CoreUnitsIndexView/components/CustomTable/list-core-unit/list-core-unit.tsx
